Rename hashing tags to iastTags
